flask_gunicorn_service/requirements.txt
Flask==2.2.3
gevent==24.2.1
gunicorn==21.2.0
Werkzeug==2.3.7
prometheus-flask-exporter==0.23.1
python-dotenv==0.21.1